♻️ refactor(View): 도메인 로직을 PrizeSummary 클래스로 분리
